Change font and font declarations
Font files were changed for more adequate files. The font for the
title or logo was changed to a static font as it's the only use I'll
make of the italic version and the variable one (the non-italic), which
is variable, was changed from ttf to woff2 for better performance.